This is modification on layout tests to ease future integration of layout implementations. Compared to previous versions, tests on non dense layout cover more cases. Layouts are filled with element coordinates. It becomes easy to check the result from `deref()`, `slice()` etc by looking at the content of the layout.

+++ tests/area/test_cuda.c
@@ -24,7 +24,7 @@ void test_device_mmap(const int device)
 	void *host_copy;
 	void *device_data;
 	struct aml_area *area;
-	int ns = sizeof(sizes) / sizeof(*sizes);
+	size_t ns = sizeof(sizes) / sizeof(*sizes);
 	int err;
 	size_t size;
 
@@ -63,7 +63,7 @@ void test_host_mmap(const int device)
 	void *host_data;
 	void *host_copy;
 	struct aml_area *area;
-	int ns = sizeof(sizes) / sizeof(*sizes);
+	size_t ns = sizeof(sizes) / sizeof(*sizes);
 	size_t size;
 
 	assert(!aml_area_cuda_create(&area, device,
@@ -91,7 +91,7 @@ void test_mapped_mmap(const int device)
 	void *host_copy;
 	void *device_data;
 	struct aml_area *area;
-	int ns = sizeof(sizes) / sizeof(*sizes);
+	size_t ns = sizeof(sizes) / sizeof(*sizes);
 	size_t size;
 	struct aml_area_cuda_mmap_options options = {.device = device,
 							.ptr = NULL, };
@@ -167,7 +167,7 @@ void test_unified_mmap(const int device)
 	void *unified_data;
 	void *host_copy;
 	struct aml_area *area;
-	int ns = sizeof(sizes) / sizeof(*sizes);
+	size_t ns = sizeof(sizes) / sizeof(*sizes);
 	size_t size;
 
 	// Data initialization
@@ -204,7 +204,7 @@ void test_unified_mmap(const int device)
 int main(void)
 {
 	int num_devices;
-	int flags;
+	unsigned int flags;
 	int has_device_map;
 	int has_unified_mem;
 	int has_register_ptr;
